+++ auth0_flutter_platform_interface/test/method_channel_credentials_manager_test.dart
@@ -31,12 +31,14 @@ void main() {
   final mocked = MockMethodCallHandler();
 
   setUp(() {
-    channel.setMockMethodCallHandler(mocked.methodCallHandler);
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, mocked.methodCallHandler);
     reset(mocked);
   });
 
   tearDown(() {
-    channel.setMockMethodCallHandler(null);
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, null);
   });
 
   group('getCredentials', () {